Include size to labels so that autolayout is improved
Autolayout "respects" a labelSize parameter if defined in a link.

There are other parameters available, such as labelPosition, and
labelOffset. Sadly, dagre (or graphlib) only allows one label
per edge. So, labels of bidirectional references get stacked one
over the other, if dagre's setLabels param is set to true. Also,
dagre overrides jointjs parameters related to labels positions,
so labels are always displayed at the center of the reference.

What I do now is give labels a size, and use jointjs label
position parameters to set them in the diagram. This label size
is respected by dagre, so extra space is left between edges for
a better display. This solution is not perfect, as jointjs
parameters come with some inconveniences.

If anything, now the position of labels can be manually set. I
will try to see if it is feasible to enable saving user changes
in the diagram.